+++ common/src/main/kotlin/net/spaceeye/vmod/events/AVSEvents.kt
@@ -5,6 +5,7 @@ import net.minecraft.server.level.ServerLevel
 import net.spaceeye.vmod.utils.SafeEventEmitter
 import org.valkyrienskies.core.api.ships.ServerShip
 import org.valkyrienskies.core.api.ships.Ship
+import org.valkyrienskies.core.impl.game.phys_entities.PhysicsEntityClient
 import org.valkyrienskies.core.impl.game.ships.ShipData
 import org.valkyrienskies.core.util.datastructures.DenseBlockPosSet
 
@@ -13,6 +14,9 @@ object AVSEvents {
     val serverShipRemoveEvent = SafeEventEmitter<ServerShipRemoveEvent>()
     val clientShipUnloadEvent = SafeEventEmitter<ClientShipUnloadEvent>()
     val splitShip = SafeEventEmitter<SplitShipEvent>() //VS event is useless
+
+    val clientPhysEntityLoad = SafeEventEmitter<ClientPhysEntityLoad>()
+    val clientPhysEntityUnload = SafeEventEmitter<Long>()
 
     data class ServerShipRemoveEvent(val ship: ShipData)
     data class ClientShipUnloadEvent(val ship: Ship?)
@@ -24,4 +28,6 @@ object AVSEvents {
         val centerBlock: BlockPos,
         val blocks: DenseBlockPosSet
     )
+
+    data class ClientPhysEntityLoad(val data: PhysicsEntityClient)
 }