+++ common/src/main/kotlin/net/spaceeye/vmod/rendering/types/special/SchemRenderer.kt
@@ -131,6 +131,8 @@ class FakeLevel(
     var offset = Vector3i(0, 0, 0)
 
     var blockEntities = mutableListOf<Pair<BlockPos, BlockEntity>>()
+    private val dummyLightEngine = SchemLightEngine()
+    private val fakeChunkSource = DummyChunkSource(this, dummyLightEngine)
 
     init {
         data.forEach { x, y, z, item ->
@@ -170,10 +172,12 @@ class FakeLevel(
     override fun playSeededSound(player: Player?, x: Double, y: Double, z: Double, sound: Holder<SoundEvent?>, source: SoundSource, volume: Float, pitch: Float, seed: Long) {}
     override fun playSeededSound(player: Player?, entity: Entity, sound: Holder<SoundEvent?>, category: SoundSource, volume: Float, pitch: Float, seed: Long) {}
 
-    private val dummyLightEngine = SchemLightEngine()
     override fun getLightEngine(): LevelLightEngine? = dummyLightEngine
     override fun getHeight(): Int = 2000000
     override fun getMinBuildHeight(): Int = -1000000
+    override fun getSectionsCount(): Int = 0
+    override fun getMinSection(): Int = 0
+    override fun getMaxSection(): Int = 0
     override fun isClientSide(): Boolean = true
     override fun blockEntityChanged(pos: BlockPos) {}
     override fun playSound(player: Player?, x: Double, y: Double, z: Double, sound: SoundEvent, category: SoundSource, volume: Float, pitch: Float) {}
@@ -192,7 +196,6 @@ class FakeLevel(
     override fun getFreeMapId(): Int = 0
     override fun getEntity(id: Int): Entity? = null
 
-    private val fakeChunkSource = DummyChunkSource(this, dummyLightEngine)
     override fun getChunkSource() = fakeChunkSource
 
     override fun getUncachedNoiseBiome(x: Int, y: Int, z: Int): Holder<Biome?>? { throw AssertionError("Shouldn't be called") }