+++ changelog.md
@@ -1,9 +1,16 @@
+# 1.6.1
+* Fixed precise placement assist side num not updating after changing preset
+* Fixed game crashing on pressing delete in Settings Preset menu without choosing preset
+* Made GravChangerMode, MassChangerMode, PhysRopeMode, ScaleMode, SensorMode, SliderMode, ThrusterMode, presettable
+* Fixed VMod not working on servers 
+
 # 1.6.0
 * Added delete button to VEntityChanger
 * Added Settings Presets
 * Fixed rare crash when raycasting
 * Rolled back schematic renderer change cuz it didn't work
 * Fixed phys bearing schem compat not working
+* Fixed possible crash when changing color in VEntity Changer
 
 # 1.5.1
 * Fixed incompat with control craft